backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
pymongo==4.6.1
redis==5.0.1
firebase-admin==7.1.0
yt-dlp>=2024.1.1
PyJWT==2.8.0
python-dotenv==1.0.0
APScheduler==3.10.4
gunicorn==21.2.0
bcrypt==4.1.2
requests==2.31.0
Werkzeug==3.0.1
imageio-ffmpeg==0.6.0

# Development dependencies
pytest==7.4.3
pytest-cov==4.1.0
black==23.12.1
flake8==7.0.0
